+++ ePunla.Common/ePunlaDB/dbo/Tables/ClaimCauses.sql
@@ -2,7 +2,7 @@ CREATE TABLE [dbo].[ClaimCauses] (
     [ClaimCauseId]    INT            IDENTITY (1, 1) NOT NULL,
     [ClaimId]         INT            NULL,
     [DamageType]      NVARCHAR (MAX) NULL,
-    [DamagedAreaSize] INT            NOT NULL,
+    [DamagedAreaSize] DECIMAL        NOT NULL,
     CONSTRAINT [PK_ClaimCauses] PRIMARY KEY CLUSTERED ([ClaimCauseId] ASC),
     CONSTRAINT [FK_ClaimCauses_Claims_ClaimId] FOREIGN KEY ([ClaimId]) REFERENCES [dbo].[Claims] ([ClaimId])
 );
